pyproject.toml
[tool.poetry]
name = "twrpdtgen"
version = "2.0.0"
description = "TWRP device tree generator. A script that generates a TWRP-compatible device tree."
license = "Apache-2.0"
readme = "README.md"
repository = "https://github.com/SebaUbuntu/TWRP-device-tree-generator"
authors = ["Sebastiano Barezzi <barezzisebastiano@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
GitPython = "^3.1.14"
Jinja2 = "^3.0.0a1"

[tool.poetry.dev-dependencies]
pytest = "^6.2.5"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
